requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
# Core training stack
# ─────────────────────────────────────────────────────────────────────────────
torch>=2.2.0
transformers>=4.40.0
datasets>=2.18.0
peft>=0.10.0
bitsandbytes>=0.43.0
accelerate>=0.29.0
safetensors>=0.4.3

# ─────────────────────────────────────────────────────────────────────────────
# Config & utilities
# ─────────────────────────────────────────────────────────────────────────────
pyyaml>=6.0.1
tqdm>=4.66.0
numpy>=1.26.0
sentencepiece>=0.2.0       # tokenizer for LLaMA / Mistral
protobuf>=4.25.0           # required by sentencepiece

# ─────────────────────────────────────────────────────────────────────────────
# Optional: experiment tracking
# ─────────────────────────────────────────────────────────────────────────────
# wandb>=0.17.0
# tensorboard>=2.16.0

# ─────────────────────────────────────────────────────────────────────────────
# Optional: faster attention (requires CUDA, build from source)
# ─────────────────────────────────────────────────────────────────────────────
# flash-attn>=2.5.0

# ─────────────────────────────────────────────────────────────────────────────
# Development & testing
# ─────────────────────────────────────────────────────────────────────────────
pytest>=8.1.0
pytest-cov>=5.0.0
